refactor: relocate notion integration to plugins folder
- moved notion package from backend/app/integrations to backend/app/plugins
- updated all internal imports, config references, and documentation
- refactored telegram bot dependency imports to resolve circularity
- added notion registration to plugins package initializer
