+++ backend/.env.example
@@ -349,13 +349,13 @@ ADMIN_EMAIL=admin@pawrrtal-ai.dev
 ADMIN_PASSWORD=admin1234
 
 # ============================================================================
-# PLUGINS  (per-workspace BYOK — declared in backend/app/integrations/*/plugin.py)
+# PLUGINS  (per-workspace BYOK — declared in backend/app/plugins/*/plugin.py)
 # ============================================================================
 # These are NOT read by Settings — each workspace's encrypted .env carries them
 # and the plugin tools read them per call via resolve_api_key(workspace_id, ...).
 # Listed here so operators know what each plugin needs to be useful:
 #
-#   Notion plugin (backend/app/integrations/notion/) — 18 tools:
+#   Notion plugin (backend/app/plugins/notion/) — 18 tools:
 #     NOTION_API_KEY            (required; get from https://www.notion.so/profile/integrations)
 #
 # To wire BYOK keys without going through the UI, POST them to
