+++ backend/app/integrations/__init__.py
@@ -5,15 +5,4 @@ iMessage as follow-ups). Adapters are responsible for the
 provider-specific I/O — translating inbound updates into Pawrrtal's
 domain calls and rendering Pawrrtal's outbound text into the provider's
 native primitives.
-
-Plugin-style integrations (Notion, future Slack-as-tools, etc.) also
-live here.  Importing this module triggers each plugin subpackage's
-import, which in turn registers the plugin against
-``app.core.plugins.registry`` so :func:`app.core.agent_tools.build_agent_tools`
-sees it on the next chat turn.
 """
-
-# Triggers Notion plugin registration as an import side-effect.  Kept
-# at module scope so a single ``import app.integrations`` is enough to
-# activate every in-tree plugin.
-from app.integrations import notion  # noqa: F401
