+++ backend/app/integrations/telegram/handlers.py
@@ -46,6 +46,7 @@ from app.crud.channel import (
 from app.integrations.telegram._attachments import (
     collect_attachments as collect_attachments,  # noqa: PLC0414
 )
+from app.integrations.telegram.bot_permissions import build_telegram_permission_check  # noqa: F401
 from app.integrations.telegram.dev_admin import resolve_or_autolink_telegram_user
 from app.integrations.telegram.model_defaults import resolve_effective_model_id
 from app.integrations.telegram.sender import TelegramSender as TelegramSender  # noqa: PLC0414
